requirements.txt
# Border Crossings Traffic Dataset Pipeline - Dependencies

# Azure SDK
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Configuration
python-dotenv>=1.0.0

# Computer Vision & Deep Learning
ultralytics>=8.0.0  # YOLO
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0

# Utilities
tqdm>=4.66.0  # Progress bars

# Development & Testing (optional, uncomment if needed)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
